requirements.txt
asttokens==2.2.1
autopep8==2.0.2
backcall==0.2.0
comm==0.1.3
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
greenlet==2.0.2
install==1.3.5
ipykernel==6.22.0
ipython==8.11.0
jedi==0.18.2
jupyter_client==8.1.0
jupyter_core==5.3.0
matplotlib-inline==0.1.6
mysql-connector-python==8.0.32
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
pika==1.3.1
pika-stubs==0.1.3
platformdirs==3.1.1
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.14.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-geoip==1.2
python-geoip-geolite2==2015.303
pytz==2022.7.1
pyzmq==25.0.2
six==1.16.0
SQLAlchemy==2.0.7
sqlalchemy-migrate==0.13.0
sqlparse==0.4.3
stack-data==0.6.2
Tempita==0.5.2
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
wcwidth==0.2.6
